modify path to components
