+++ docs/reference/components/component_markdowndoc.md
@@ -1,0 +1,8 @@
+# Reference for `MarkdownDoc`
+
+The ``MarkdownDoc`` is the entity returned by ``chunknorris``'s parsers. It's purpose it mainly to be fed to the ``MarkdownChunker``.
+
+::: chunknorris.core.components.MarkdownDoc
+    handler: python
+    options:
+      show_source: false
