+++ docs/reference/components/component_markdownline.md
@@ -1,0 +1,8 @@
+# Reference for `MarkdownLine`
+
+The ``MarkdownLine`` represents a markdown-formatted line and some associated features (whether the line is a header, or belongs to a code block for example).
+
+::: chunknorris.core.components.MarkdownLine
+    handler: python
+    options:
+      show_source: false
