+++ mkdocs.yml
@@ -63,6 +63,10 @@ nav:
           #- "Implement my own pipeline": examples/example1.ipynb # TODO
   - Reference:
       - "API reference": reference/index.md
+      - Core components:
+          - Chunk: reference/components/component_chunk.md
+          - MarkdownDoc: reference/components/component_markdowndoc.md
+          - MarkdownLine: reference/components/component_markdownline.md
       - Chunkers:
           - MarkdownChunker: reference/chunkers/chunker_markdown.md
       - Parsers:
