+++ src/chunknorris/chunkers/markdown_chunker.py
@@ -1,9 +1,8 @@
 from typing import Literal
 
+from ..core.components import Chunk, MarkdownDoc, MarkdownLine, TocTree
 from ..decorators.decorators import timeit, validate_args
-from ..parsers.markdown.components import MarkdownDoc, MarkdownLine
 from .abstract_chunker import AbstractChunker
-from .tools.tools import Chunk, TocTree
 
 
 class MarkdownChunker(AbstractChunker):
