--- src/chunknorris/chunkers/tools/__init__.py
@@ -1,1 +1,0 @@
-from .tools import Chunk, TocTree
