+++ src/chunknorris/parsers/notebook/jupyter_notebook_parser.py
@@ -5,8 +5,8 @@ from typing import Any
 import nbformat
 from pydantic import ValidationError
 
+from ...core.components import MarkdownDoc
 from ..abstract_parser import AbstractParser
-from ..markdown.components import MarkdownDoc
 from .schemas import JupyterNotebookContent
