+++ src/chunknorris/pipelines/abstract_pipeline.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 
 from ..chunkers.abstract_chunker import AbstractChunker
-from ..chunkers.tools import Chunk
+from ..core.components import Chunk
 from ..parsers.abstract_parser import AbstractParser
