+++ tests/chunknorris/chunkers/test_markdown_chunker.py
@@ -1,6 +1,5 @@
 from chunknorris.chunkers import MarkdownChunker
-from chunknorris.chunkers.tools import Chunk
-from chunknorris.parsers.markdown.components import MarkdownDoc
+from chunknorris.core.components import Chunk, MarkdownDoc
 
 
 def test_chunk_string(
