+++ tests/chunknorris/parsers/test_html_parser.py
@@ -1,5 +1,5 @@
 from chunknorris.parsers import HTMLParser
-from chunknorris.parsers.markdown.components import MarkdownDoc
+from chunknorris.core.components import MarkdownDoc
 
 
 def test_parse_string(
