+++ tests/chunknorris/parsers/test_markdown_parser.py
@@ -1,5 +1,5 @@
 from chunknorris.parsers import MarkdownParser
-from chunknorris.parsers.markdown.components import MarkdownDoc
+from chunknorris.core.components import MarkdownDoc
 
 
 def test_markdowndoc(md_strings_in: list[str]):
