+++ tests/chunknorris/parsers/test_pdf_parser.py
@@ -3,7 +3,7 @@ import re
 import pymupdf  # type: ignore -> no stubs
 
 from chunknorris.parsers import PdfParser
-from chunknorris.parsers.markdown.components import MarkdownDoc
+from chunknorris.core.components import MarkdownDoc
 
 
 def test_parse_file(pdf_parser: PdfParser, pdf_filepath: str):
