+++ tests/chunknorris/types/test_types.py
@@ -1,4 +1,4 @@
-from chunknorris.parsers.markdown.components import MarkdownDoc
+from chunknorris.core.components import MarkdownDoc
 
 
 def test_markdowndoc(md_strings_in: list[str]):
